--- pony/package.json
@@ -1,5 +1,0 @@
-{
-  "main": "../dist/pony.cjs",
-  "module": "../dist/pony.js",
-  "types": "../dist/types.d.ts"
-}
